idkkkkk

+++ Backend/src/main/java/cs309/backend/jpa/repo/MessageRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface MessageRepository extends JpaRepository<MessageEntity, Long> {
+public interface MessageRepository extends JpaRepository<MessageEntity, Integer> {
 
     @Procedure(name = "saveMessage")
     void saveMessage(
